MOdified class names, reorganized files better
